requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
joblib==1.5.3
numpy==2.4.4
psycopg2-binary==2.9.12
PyJWT==2.12.1
scikit-learn==1.8.0
scipy==1.17.1
sqlparse==0.5.5
threadpoolctl==3.6.0
